+++ src/GridSolutions.jl
@@ -9,8 +9,12 @@ import PowerDynBase: GridDynamics, internalindex
 "Abstract super type for all data structures representing solutions of power grid models."
 abstract type AbstractGridSolution end
 
+abstract type AbstractSingleGridSolutions <: AbstractGridSolution end
+
+ # to ensure broadcasting can be used
+ @inline Broadcast.broadcastable(sol::AbstractSingleGridSolutions) = Ref(sol)
 """
-    struct GridSolution <: AbstractGridSolution
+    struct GridSolution <: AbstractSingleGridSolutions
         dqsol::AbstractTimeseriesSolution
         griddynamics::GridDynamics
     end
@@ -41,7 +45,7 @@ plot(sol, n, s, plots_kwargs...)
 ```
 where `n` and `s` are as in the accessing of plots, and `plots_kwargs` are the keyword arguments for Plots.jl.
 """
-struct GridSolution <: AbstractGridSolution
+struct GridSolution <: AbstractSingleGridSolutions
     dqsol::AbstractTimeseriesSolution
     griddynamics::GridDynamics
     function GridSolution(dqsol::AbstractTimeseriesSolution, griddynamics::GridDynamics)
@@ -56,12 +60,19 @@ TimeSeries(sol::GridSolution) = sol.dqsol
 tspan(sol::GridSolution) = (TimeSeries(sol).t[1], TimeSeries(sol).t[end])
 tspan(sol::GridSolution, tres) = range(TimeSeries(sol).t[1], stop=TimeSeries(sol).t[end], length=tres)
 
-(sol::GridSolution)(t, ::Colon, sym::Symbol, args...) = sol(t, eachindex(Nodes(sol)), sym, args...)
+(sol::GridSolution)(t, ::Colon, sym::Symbol, args...; kwargs...) =
+    sol(t, eachindex(Nodes(sol)), sym, args...; kwargs...)
 (sol::GridSolution)(t, n, sym::Symbol, args...) = begin
-    if ~all( 1 .<= n .<= length(Nodes(sol))  )
+    if ~all( 1 .<= n .<= length(Nodes(sol)) )
         throw(BoundsError(sol, n))
+    else
+        sol(t, n, Val{sym}, args...)
     end
-    sol(t, n, Val{sym}, args...)
+end
+(sol::GridSolution)(t::Number, n::Number, ::Type{Val{:u}}) = begin
+    u_real = TimeSeries(sol)(t, idxs= 2 * n - 1)
+    u_imag = TimeSeries(sol)(t, idxs= 2 * n)
+    u_real + im * u_imag
 end
 (sol::GridSolution)(t, n, ::Type{Val{:u}}) = begin
     u_real = @>> TimeSeries(sol)(t, idxs= 2 .* n .- 1) convert(Array)
@@ -77,33 +88,36 @@ end
 (sol::GridSolution)(t, n, ::Type{Val{:p}}) = sol(t, n, :s) .|> real
 (sol::GridSolution)(t, n, ::Type{Val{:q}}) = sol(t, n, :s) .|> imag
 (sol::GridSolution)(t, n, ::Type{Val{:int}}, i) = @>> TimeSeries(sol)(t, idxs=internalindex(sol, n, i)) convert(Array)
+(sol::GridSolution)(t::Number, n::Number, ::Type{Val{:int}}, i::S) where {S <: Union{Number,Symbol}} = TimeSeries(sol)(t, idxs=internalindex(sol, n, i))
 (sol::GridSolution)(t, n, ::Type{Val{sym}}) where sym = sol(t, n, Val{:int}, sym)
-
 
 # define the plotting recipes
 using RecipesBase
 
 "Create the standard variable labels for power grid plots."
 tslabel(sym, node) = "$(sym)$(node)"
-tslabel(sym, n::AbstractArray) = tslabel.(Ref(sym), n)
-tslabel(sym, node, i) = "$(sym)$(node)_$(i)"
-tslabel(sym, n::AbstractArray, i) = tslabel.(Ref(sym), n, Ref(i))
 "Transform the array output from DifferentialEquations.jl correctly to be used in Plots.jl's recipes."
 tstransform(arr::AbstractArray{T, 1}) where T = arr
 tstransform(arr::AbstractArray{T, 2}) where T = arr'
 
-const PLOT_TTIME_RESOLUTION = 10_000
+const PLOT_TTIME_RESOLUTION = 1_000
 
-@recipe function f(sol::GridSolution, ::Colon, sym::Symbol, args...)
+@recipe function f(sol::AbstractSingleGridSolutions, ::Colon, sym::Symbol, args...)
     sol, eachindex(Nodes(sol)), sym, args...
 end
-@recipe function f(sol::GridSolution, n, sym::Symbol, args...; tres = PLOT_TTIME_RESOLUTION)
+
+function getPlottingData(sol::AbstractSingleGridSolutions, n, sym::Symbol, args...; tres)
+    t = tspan(sol, tres)
+    t, tstransform(sol(t, n, sym, args...))
+end
+
+@recipe function f(sol::AbstractSingleGridSolutions, n, sym::Symbol, args...; tres = PLOT_TTIME_RESOLUTION)
     if sym == :int
-        label --> tslabel(sym, n, args[1])
-    else
-        label --> tslabel(sym, n)
+        throw(PowerDynamicsPlottingError("deprecated usage of `:int` for indexing the internal variables: cannot be used for plotting. Use the symbol representing the internal variables instead, e.g. `:ω` for `SwingEq`."))
     end
+    label --> tslabel.(sym, n)
     xlabel --> "t"
     t = tspan(sol, tres)
     t, tstransform(sol(t, n, sym, args...))
+    getPlottingData(sol, n, sym, args...; tres = tres)
 end